march_mazeness-break_status
working on wall detection. still need to do fog/walls. fog possibly
could be done by saving the line1-line20 as "default" and having a for
loop that checks if an item is not adjacent and then replaces with a 0.
other possible route is just getting rid of all elements not in the
directly adjacent boxes.